+++ Corona-Warn-App/src/main/java/de/rki/coronawarnapp/eventregistration/events/TraceLocationCreator.kt
@@ -2,7 +2,7 @@ package de.rki.coronawarnapp.eventregistration.events
 
 import de.rki.coronawarnapp.environment.EnvironmentSetup
 import de.rki.coronawarnapp.presencetracing.checkins.qrcode.TraceLocation
-import de.rki.coronawarnapp.eventregistration.storage.repo.TraceLocationRepository
+import de.rki.coronawarnapp.presencetracing.storage.repo.TraceLocationRepository
 import okio.ByteString
 import okio.ByteString.Companion.toByteString
 import java.security.SecureRandom
