+++ Corona-Warn-App/src/main/java/de/rki/coronawarnapp/util/di/ApplicationComponent.kt
@@ -17,7 +17,7 @@ import de.rki.coronawarnapp.diagnosiskeys.DiagnosisKeysModule
 import de.rki.coronawarnapp.diagnosiskeys.DownloadDiagnosisKeysTaskModule
 import de.rki.coronawarnapp.diagnosiskeys.storage.KeyCacheRepository
 import de.rki.coronawarnapp.environment.EnvironmentModule
-import de.rki.coronawarnapp.eventregistration.EventRegistrationModule
+import de.rki.coronawarnapp.presencetracing.PresenceTracingModule
 import de.rki.coronawarnapp.http.HttpModule
 import de.rki.coronawarnapp.nearby.ENFClient
 import de.rki.coronawarnapp.nearby.ENFModule
@@ -77,7 +77,7 @@ import javax.inject.Singleton
         StatisticsModule::class,
         DataDonationModule::class,
         SecurityModule::class,
-        EventRegistrationModule::class,
+        PresenceTracingModule::class,
         CoronaTestModule::class,
     ]
 )
