+++ Corona-Warn-App/src/test/java/de/rki/coronawarnapp/eventregistration/events/TraceLocationMapperTest.kt
@@ -3,7 +3,7 @@ package de.rki.coronawarnapp.eventregistration.events
 import de.rki.coronawarnapp.presencetracing.checkins.qrcode.TraceLocation
 import de.rki.coronawarnapp.presencetracing.checkins.qrcode.toTraceLocation
 import de.rki.coronawarnapp.presencetracing.checkins.qrcode.toTraceLocations
-import de.rki.coronawarnapp.eventregistration.storage.entity.TraceLocationEntity
+import de.rki.coronawarnapp.presencetracing.storage.entity.TraceLocationEntity
 import de.rki.coronawarnapp.server.protocols.internal.pt.TraceLocationOuterClass.TraceLocationType.LOCATION_TYPE_PERMANENT_OTHER
 import de.rki.coronawarnapp.server.protocols.internal.pt.TraceLocationOuterClass.TraceLocationType.LOCATION_TYPE_TEMPORARY_OTHER
 import io.kotest.matchers.shouldBe
